Separate internal and external auth controllers
